feat: `CamerasConfig` works ininspector

+++ src/blocks.rs
@@ -3,7 +3,7 @@ use crate::prelude::*;
 /// The unique identifier for a persistent block in the world
 #[derive(Reflect, Debug, Clone, Copy, Component, Serialize, Deserialize, PartialEq, Eq)]
 #[reflect(Component)]
-pub struct BlockId(u128);
+pub struct BlockId(u64);
 
 impl Default for BlockId {
 	fn default() -> Self {
